backend/requirements.txt
# fastapi
# uvicorn
pydantic==1.10.13
langserve[all]
langchain
openai
python-dotenv